# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Civil Wars API
  version: 1.0.0
  description: Maschinenlesbare API-Spezifikation fuer Civil Wars (Backend).
servers:
  - url: http://localhost:3000
    description: Lokale Entwicklung
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Benutzer registrieren
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registrierung erfolgreich
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '400':
          description: Validierungs- oder Konfliktfehler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '503':
          description: Kein freier Koordinatenplatz verfuegbar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /auth/login:
    post:
      tags: [Auth]
      summary: Benutzer einloggen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login erfolgreich
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '400':
          description: Validierungsfehler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Ungueltige Credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Account inaktiv
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Access-Token via Refresh-Token erneuern (Rotation)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token erfolgreich erneuert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshSuccessResponse'
        '400':
          description: Validierungsfehler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Refresh-Token ungueltig oder abgelaufen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Account inaktiv
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /auth/me:
    get:
      tags: [Auth]
      summary: Eigenes User-Profil aus Token lesen
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Userdaten
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/AuthUser'
        '401':
          description: Kein oder ungueltiger Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: User nicht gefunden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /me:
    get:
      tags: [Me]
      summary: Vollstaendiger Spielerstatus
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Spielerstatus
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerStatus'
        '401':
          description: Kein oder ungueltiger Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
    AuthSuccessResponse:
      type: object
      properties:
        message:
          type: string
        token:
          type: string
        refresh_token:
          type: string
        user:
          $ref: '#/components/schemas/AuthUser'
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          minLength: 20
    RefreshSuccessResponse:
      type: object
      properties:
        message:
          type: string
        token:
          type: string
        refresh_token:
          type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ValidationErrorItem:
      type: object
      properties:
        code:
          type: string
        path:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
        message:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorItem'
    ResourceValues:
      type: object
      properties:
        geld:
          type: number
        stein:
          type: number
        stahl:
          type: number
        treibstoff:
          type: number
        strom:
          type: number
...
    Building:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        anzahl:
          type: integer
        level:
          type: integer
...
    QueueEntry:
      type: object
      properties:
        id:
          type: integer
        building_type_id:
          type: integer
        anzahl:
          type: integer
        fertig_am:
          type: string
          format: date-time
...
    PlayerStatus:
      type: object
      properties:
        resources:
          $ref: '#/components/schemas/ResourceValues'
        buildings:
          type: array
          items:
            $ref: '#/components/schemas/Building'
        queue:
          type: array
          items:
            $ref: '#/components/schemas/QueueEntry'
        strom:
          type: object
          additionalProperties: true
        production:
          type: object
          additionalProperties: true
...